playbooks/roles/validator_service/requirements.txt
flask>=3.1.0
boto3>=1.37.0
python-dotenv>=1.1.0